implement review-feedback (error in test_plot_raw_traces still unsolved)
